Add activities to database lib
